requirements.txt
absl-py==2.1.0
android_env
dm_env==1.6
fuzzywuzzy==0.18.0
google-generativeai==0.5.1
grpcio-tools
immutabledict==2.0.0
IPython
jsonschema==4.17.3
matplotlib==3.6.1
numpy==1.26.3
openai
opencv-python
pandas==2.1.4
pydub
requests==2.31.0
tenacity
termcolor